refactor methods to smaller packages
