+++ ic-asset/src/asset_canister/method_names.rs
@@ -1,0 +1,4 @@
+pub(crate) const CREATE_BATCH: &str = "create_batch";
+pub(crate) const CREATE_CHUNK: &str = "create_chunk";
+pub(crate) const COMMIT_BATCH: &str = "commit_batch";
+pub(crate) const LIST: &str = "list";
