+++ ic-asset/src/asset_canister/mod.rs
@@ -1,1 +1,4 @@
 pub(crate) mod protocol;
+pub(crate) mod batch;
+pub(crate) mod method_names;
+pub(crate) mod chunk;
