+++ ic-asset/src/lib.rs
@@ -2,5 +2,6 @@ mod asset_canister;
 mod content;
 mod content_encoder;
 mod sync;
+mod convenience;
 
 pub use sync::sync;
